requirements.txt
# Core Streamlit
streamlit>=1.28.0

# LangChain and related packages
langchain>=0.1.0
langchain-core>=0.1.14
langchain-community>=0.0.34
langchain-google-genai

# Text splitting / chunking
langchain-text-splitters>=0.1.0

# Vector databases and embeddings
faiss-cpu>=1.7.4

# Environment and utilities
python-dotenv>=1.0.0

# Dataset handling + embeddings
pandas>=2.0.0
sentence-transformers>=2.2.0

# Text processing
chardet>=5.2.0  # For better encoding detection